feat(editor): editor object in place of store exports
