+++ packages/editor/src/core/index.ts
@@ -1,6 +1,6 @@
 export { SubDocument } from './sub-document'
 export type { SubDocumentProps } from './sub-document'
 export { Editor } from './editor'
-export type { EditorProps } from './editor'
+export type { EditorProps, EditorData } from './types'
 export { PreferenceContext } from './contexts'
 export type { Preference } from './contexts'
